doc/cr: improve variables declarations coding rules
The coding rules about variables declarations were dubious on 2 points:

* We allowed to declare 2 scalars on the same line, like `int a, b` but
  not 2 structures like `foo_t a, b` which doesn't make sense.
* We allowed to declare 2 non-pointer variable on the same line, like
  `int a, b` but not 2 pointers like `int *a, *b` when what must be
  discouraged is the mix of pointers and non-pointers variables.

This new version of the CR are more coherent and also less strict. In
several cases, it's meaningful to have several variable declared on the
same line, even if they are pointers.

Of course most of the time it's preferable to keep one declaration per
line, if only to encourage per-variable documentation. And this is
emphasized in the CR.

Change-Id: I8929b6e720d01b96f384fc4a60124a218fcde3eb
Priv-Id: cba08da9bad2bf4ee3cab6f8762b6e1c076d319a