qps/hat: fix broken test in 'remove' function
In the function that removes entries from a qhat, the QHAT stats were
badly updated for entries removed for flat leafs: the test checking if
the removed entry had a null value (which means that it is absent) was
broken for 128 bits values and totally inverted for smaller values.

This fix should not have such an impact as stats are deactivated most of
the time and have no functional impact when they are.

Change-Id: I3fd86c833f195902984a6df0c63973ac8e2427ef
Priv-Id: c910b87a3eed5b9dd3543c3f77d98eea2a2063c5