qps/hat: fix hole in struct 'qhat_tree_enumerator_t'
It seems like some effort have been involved in order to make the
structure compact but there was a hole in it.

This patch moves the attribute 'value_tab' in order to fix that.

Summary by pahole, before:

    /* size: 80, cachelines: 2, members: 10 */
    /* sum members: 72, holes: 2, sum holes: 8 */
    /* last cacheline: 16 bytes */

After:

    /* size: 72, cachelines: 2, members: 10 */
    /* last cacheline: 8 bytes */

No risk for backward-compatibility: the enumerator is not part of the
stuff written to disk in QPS.

Change-Id: I55a13d2454b8d5b8800af6c68f47089d2281a0b2
Priv-Id: 571c8a8fbd7546f847d4f4719c3479b9a12d6f9f