net/http: fix http2 client inactivity bug
Fixed HTTP/2 client inactivity bug that results in HTTPC_STATUS_ABORT.
The bug is caused by the HTTP/2 being unaware of a written httpc query
reply while the http2 conn is waiting for POLLIN event from the socket
because its ob was empty when it set its fd mask. The solution is to
catch signaled writes to the httpc_t's ob during httpc_set_mask and
signal them to the underlying http2_conn_t object.

Change-Id: If3e9a546d7686356ba24ab4e0405e5c808602dfc
Priv-Id: 31167269ca121ea6a0e30aa05c84efdf5bfbafc1