z: add some colors when we print diff between IOPs
When Z_ASSERT_IOPEQUAL() (or similar macro) fails, we print a YAML diff
in the output report so that the user can get a good understanding of
how the test result differs from the expectations.

This patch adds some colors in the diff output (only when the tests are
run in a terminal that supports it).

Change-Id: I3a8f2bded75ae01dbbee0549c37df2b4f0135fee
Priv-Id: a4065d14e14303df62aee506f0e5afe55d5cc88e