qps/hat: fix typo in 'qhat_split_leaf()'
Typo: PATH_MAX was used instead of QHAT_MAX_DEPTH, making a case where
we want to "flatten" a QHAT leaf unreachable.

The problem was present from the introduction of flat leafs in
85ceb9d544743fb8f3870279ae963a292e09521e.

This patch fixes the typo but does not change the behavior: we must get
a better understanding of the case and build a NR test first, otherwise
it would be like accepting the risk of letting a fossil disease escape
from an recently unfrozen mammoth. A FIXME has been left next to the
neutralized code.

Seen while trying to change the type of attribute qhat_path_t::depth
from 'int' to 'uint8_t': the value of PATH_MAX - 1 did not fit anymore.

Change-Id: I8a0cfbad8fc4430c741597c229def473c04385dd
Priv-Id: d11543ee192743224c52538033e6129fdef9ebfa