net/http: http2: fix count httpd towards the limit
Fixed a bug related to not counting the http2-transported httpds towards
the limit of connections set in httpd_cfg_t object (nb_conns). This
limit is decremented by one in httpd_wipe but not incremented when httpd
is spawned as an http2 stream. This can lead to stopping accepting
connections because of the test in httpd_on_accept
    if (cfg->nb_conns >= cfg->max_conns) {
            close(sock);
    }
since the nb_conns became UINT_MAX due to underflow.

Change-Id: I303c5b6699763f17559a1bfcee7adc74861577f0
Priv-Id: 123854465202003e1442b373bd8b2eb163b56b5f