iop/iop.blk: Add typedefs in package, link them to their ref objects
* What
This commit concerns the handling of IOP typedefs when parsing IOP
files so typedef objects can be registered and linked to their
aliased objects.

* Why
The introduction of IOP typedefs allows more flexible and reusable
data structures in the parsing logic.

* How
By adding a new IOP_OBJ_TYPE_TYPEDEF and a pointer to iop_typedef_t
in the desc structure.

By implementing the registration of the typedef in
iop_register_package_env() to make it accessible in the environment.

By modifying iop_get_class_by_fullname() in order to handle typedefs as
aliases to structs.

Change-Id: Ia08298993d8b24a3805e32898b2af968ba20aa62
Priv-Id: 15499eaeed4fc96e679890419e73fdae9e6dee18