qps/hat: simplify enumerator code (drop "light resync" code)
When a QPS hat is modified in such extent that a node is added or
removed, we increment a "generation" that allows us to avoid messing up
the memory when performing QHAT modifications while scanning it: the
enumerator remembers the "generation" of the QHAT and resync if it
changed under its feet.

The code seems to work OK this way but there is a gray zone that appears
when the QHAT was changed (entries inserted or removed) with no node
added or removed: we used to "resync" the enumerator with specific code
supposedly lighter than the "big resync" code.

This patch eliminates the "gray zone" and removes the "light resync"
code. The reason is simple: we want to be KISS and I do not think that
the "light resync" had a sufficiently significant impact on the
performance when compared to the "full resync".

In order to do so, we increment the "struct_gen" of the QHAT after any
insertion or removal into a "compact" leaf because it performs
insertions and removals in arrays (keys and values arrays) and the
resulting memmoves can change the positions of several entries, so
that we can drop all the code that fixup the enumerator when detecting
such change (the gray zone). There is no such impact in case of change
in a flat leaf as key positions cannot change.

Change-Id: Id8ea39d9c7768c2f790052b3b2f46bc94680e74a
Priv-Id: d38602091552caab2d48cf6a0dfb6e26d0c1e965