qps/hat: fix behavior inconsistency in enumerator
When using 'qhat_tree_enumerator_go_to()' or
'qhat_tree_enumerator_next()' we had two different behaviors depending
on the type of QHAT leaf the enumerator was scanning:

* If the leaf was compact, 'go_to' and 'next' could not stop at a
  missing entry, it went directly to the first non-missing key first.

* If the leaf was flat, both functions would stop even on an missing
  entry.

The second behavior is not what we would expect from container
enumerator. A basic example about why it is bad is the following:

 count = 0;
 qhat_for_each_unsafe(en, &trie) {
     count++;
 }

After these few lines, "count" may be bigger than the number of entries
in the QHAT, which is quite hard to accept.

Also fix documentation about attribute 'qhat_tree_enumerator_t::count':
it is used for both compact and flat leaves.

Change-Id: Ic7288c636cf48e2ebe743c0bda4f18db5ebfb200
Priv-Id: 5f68c00f24e44a23de957fba920860640d4d6363