net/http: add a comment on httpc_pool_get
The function httpc_pool_get is likely to return NULL as there may be no
available httpc_t when calling it.
A common practice is to listen on `on_ready` callback to be notified
when a httpc_t becomes ready after an unsuccessful call to
httpc_pool_get.
Comment the function httpc_pool_get to explain this behavior.

Change-Id: Id5ed26623e44dd6d733dd1873bdcdfbc3821a098
Priv-Id: 173666198f726306b3c752cc6cac93ac5fd62367