http: fix sending of GOAWAY at shutdown
in httpc_close_http2_pool, want_write was not set, so POLLOUT was not
set, and we did not go to http2_conn_on_event.

In http2_conn_on_event, prevent sending goaway multiple times by
checking w->is_shutdown_sent

It looks like GOAWAY is not treated correctly on server side, so this
patch is not enough for proper shutdown, but at least clients now send a
GOAWAY frame at shutdown.

Change-Id: I00f69221946b7a28c658197de757faefd0495916
Priv-Id: ccfa9f82eba30f0fb01e7dc05d322bea17b475f7