ichttp: add callback on exception.
When an exception is raised, the HTTP error code returned is always 500
(internal server error).

In some cases we want to convert the platform's internal error code into
an HTTP error code.

We have therefore added an optional callback which, for a given
exception, modifies the HTTP error code returned.

Change-Id: I463db3aedffed3eccc189cdc5d541fbc70c4ef44
Priv-Id: 09a6529a22de77a2ec013f58a89946dfce9ebead