iopc/iopc-lang-c: dump field attributes for typedef.
* What
Dump the field attributes associated to the typedef in iop_typedef_t.
Improve the documentation of iop_typedef_t.

* Why
This way, it is possible to do some introspection on the attributes of a
typedef.

* How
Like the fields of structs or unions, use iopc_dump_field_attrs() and
iopc_dump_field_attrs_content_src() to dump the fields attributes of the
typedef.
Adapt the code generation tests with the new attributes.

Change-Id: I7fa6555cd7401a832160e352f6e0e6b36d0d8b72
Priv-Id: a85a400bc31cad400631fd8542b1095355f23942