iop/dso: fix backward compat typedef.
* What
Do not duplicate the packages anymore to ensure the backward
compatibility.

* Why
Trying to duplicate the packages was actually a bad idea.
This worked for packages without dependencies, but if a package had
dependencies, the references were not adapted when duplicating the
packages.
We could have a reference duplication when loading the dso,
unfortunately, there are two obstacles for this:
1. Doing it in iopdso_register_pkg() is not possible because the
   dependencies are loaded after loading the current package.
   Changing this order can have some huge impacts.
   We could do it before calling iopdso_register_pkg() in
   iop_dso_register_(), but this would not solve the second obstacles.
2. When the dso references a package that is another dso, the resolution
   is directly done by the system depending on the dso symbols.
   Trying to fight against the system and dso symbols is not something
   that is really appropriate.

* How
We now need to check the dso version before accessing the field.
This is not perfect solution because we need to be very careful when
using iop_pkg_t now.
We also need to pass the dso of the second pkg in
iop_pkg_check_backward_compat().
Ideally we should find another way to do this, but I don't know how
right now.

Change-Id: I704ac4e70b994c293337765c061e50ed42d845bb
Priv-Id: e299b5f97bb0b467d93cfc4a7da0b708353374c1