iopy: support typedef types when loading a DSO.
When loading a DSO with typedefs, create classes and metaclasses to
integrate the typedef in the hierarchy and being able to use them in
Python with IOPy.

A typedef class is a proxy class to the referenced type.

See the documentation in iopy.pyx to see how are typedef classes are
created and used.

Change-Id: I6b8b6fba30ac4af654e4765a9b563c630632bc49
Priv-Id: 04d6ac067bbf41f277d537c74b08e573267f4cec