z: fix todo-pass definition
The todo-fail/todo-pass statuses were made to simplify test driven
development. With them a developer can commit a test before implementing
the corresponding feature, or a non-regression test before actually
fixing the code.

However, both todo-fail and todo-pass were defined as non fatal
statuses. Meaning that a test flagged as Z_TODO could be easily
forgotten and left this way even after implementing the feature.

Thus we now define the todo-pass status as fatal. Either if a developer
forgot to remove the Z_TODO or fixed a failing test involuntary we need
to be sure that the change of state from todo-fail to todo-pass will be
notice.

Also note that this behavior is already implemented in z.py because
@z.ZTodo is mapped to @unittest.expectedFailure which does the same
thing.

Change-Id: If938d73ef362af780965b4aa7f4ec0623f2339aa
Priv-Id: da56b0b762202c9c933a2b0ec413612b6346e956