qpsstress: change size rule and check data pattern
* What
Provide a quick and easy way to change the allocation and reallocation
rule for stimulating qps, so we can have a better range of allocations
made in tlsf or qps page maps. Provide also a way to check data inside
QPS, based on a computed pattern from handle id.

* Why
This tool lacks all checks related to QPS storage in tlsf or QPS page
maps, so we provide a way to check that operations don't change the
integrity of all QPS data.

* How
By implementing check pattern mechanism after each qpsstress operation
and by performing a wilder range for the size allocated for a handle
(before, QPS page maps were never stimulated because memory allocated
for handles could not exceed 4096 bytes).

Change-Id: If7b5371ebbc4a98b45098a0674263e1be126812f
Priv-Id: deaf960d898d95d7915baaeddff6cf74c2f50675