qps: qps object correlation
* What
This patch processes qps objects and checks that usage of these objects
is referenced once (and only once). Any handle or qps page which is
faulty is caught in qps_dissect_process_object_usage() function. As this
operation is memory consuming, the operator can disable correlation on
demand.

* Why
As part of qps dissect analysis tool, we want to correlate usage of
handles/qps pages in QKV against blocks physically allocated in qps.
Doing that allow us to track remaining handles not yet analyzed by qps
dissect tool.

* How
By doing the next operations:
 - Register all QPS allocated objects from tlsf and qps page maps (a qps
   page map can actually be a container of qps pointers for handles in
   tlsf maps)
 - Register usage of QPS objects (through a parsing of a QKV shard which
   uses QPS, for example)
 - Process meta data of these QPS objects and check for potential issues

At the end of these operations, we are able to track remaining handles
not seen yet.

Change-Id: Ieedd902b99e62bcb12e1a229bbe221f4288253dd
Priv-Id: d6811f7ddd3c2ad54bdb449f837c03ccc1d7ae1a