iop: add a function to help using 'iop_value_from_field()'
When using 'iop_value_from_field()', it sets a field in an
'iop_value_t'. In code that uses that function in a context of IOP
introspection, we usually do a 'switch' on the IOP type of the field to
know which field to read, but there are a lot more cases than there are
fields we really want to read. As a consequence, the 'switch' code is
way bigger than it should be:

    switch (field_type) {
    case IOP_T_BOOL:
        avg_acc->sum += value.b;
        break;

    case IOP_T_I8:
    case IOP_T_I16:
    case IOP_T_I32:
    case IOP_T_I64:
    case IOP_T_ENUM:
        avg_acc->sum += value.i;
        break;

    case IOP_T_U8:
    case IOP_T_U16:
    case IOP_T_U32:
    case IOP_T_U64:
        avg_acc->sum += value.u;
        break;

    case IOP_T_DOUBLE:
        avg_acc->sum += value.d;
        break;

    case IOP_T_STRING:
    case IOP_T_XML:
    case IOP_T_DATA:
    case IOP_T_VOID:
    case IOP_T_STRUCT:
    case IOP_T_UNION:
        /* Totally unexpected. */
        assert(false);
        break;
    }

With that new function, the code above can be written this way:

    switch (iop_value_field_from_type(field_type)) {
    case IOP_VALUE_I:
        avg_acc->sum += value.i;
        break;

    case IOP_VALUE_U:
        avg_acc->sum += value.u;
        break;

    case IOP_VALUE_D:
        avg_acc->sum += value.d;
        break;

    case IOP_VALUE_NONE:
    case IOP_VALUE_S:
    case IOP_VALUE_PTR:
        /* Totally unexpected. */
        assert(false);
        break;
    }

Please note that the boolean fields are now written using the 'u' field.
Of course, it can still be read using the 'b' field but we won't trigger
an invalid read anymore if we read it using the 'u' field.

Change-Id: I950998a8e4321b1dfebd1fe0d04ff3bd39791eb9
Priv-Id: 2055e96e080b2b22c8ab760930cceb22a460fe42