qpsstress: replay sequence from a previous run
* What
Provide an easy way to reproduce thousand of operations from an initial
qpsstress log (provided in parameter) without rr tool. This mechanism
allow also us to factorize and trigger problematic sequences without the
need to hardcode steps in the tool itself.

* Why
Debugging thousand of operations with rr seems to be problematic on this
tool and we don't have a proper way to manage our own sequences from a
simple text file to investigate bugs.

* How
By implementing replay mechanism from a previous qpsstress log and use
the log patterns used in this tool to replay/create sequences.

Change-Id: I8bdfdf5dcadffc691cf5f331da3f4a51a5bea662
Priv-Id: 6d3170c8b9a838c529f036bd4476d69280c513c4