waftools: rework and fix compilation database with clang.
Multiple rework and fixes in this commit for the compilation database:

1.
In order to generate the compilation database, we iterate through the
list of C, C++, Blk and Blkk tasks to retrieve the list of commands used
to compile the source files.
Unfortunately, since our standard compiler is GCC and not Clang, for
plain C and C++ source files, we retrieved commands to compile the
source files with GCC and not Clang.

The issue is that the JSON compilation database is very closely related
to Clang: https://clang.llvm.org/docs/JSONCompilationDatabase.html.
So using the arguments for GCC is kind of wrong and lead to unrecognized
arguments. A filtering system was already in-place, but it was not
perfect and not really flexible.

The solution is, when generating the compilation database, to force
using Clang just for the compilation database instead of GCC.
This way we are sure that the commands used for the tasks are well
designed for Clang.

2.
Fix using 'command' instead of 'arguments' if we have the task command
as a string and not as a list of arguments.
This is well specified in the spec:
https://clang.llvm.org/docs/JSONCompilationDatabase.html

3.
Improve the code a bit by retrieving the waf task classes only once and
not when iterating on the task generators.

4.
Generate compilation commands for headers.
Instead of only generate the compilation commands for source files,
also generate them for the headers.
This way, editors can also provide warnings on header files.
'-xc-header' and '-xcxx-header' are used for C and C++ headers, while
'-xc' and '-xcxx' are used for C and C++ sources (useful for .blk and
.blkk files).
Unfortunately, due to a bug in clang < 16 with '-fsyntax-only',
'-Wno-unused' is also required for header types.

5.
Also add the generated files in the compilation database as it can be
useful in some cases.

Change-Id: I07782e5d49ebc178d3ef7878f48fa1d9afa5bfe3
Priv-Id: 10b80e254301b94c291e4078f14fb3baa7bd94d3