compile_commands: fix compilation database for external command usage.
External commands like Doxygen or xunused can use Clang with
`clang_parseTranslationUnit()` by passing the
source files not with `source_filename`, but in `unsaved_files`.

The consequence is that in the list of arguments, we need to be sure
that the exact absolute path of source file is used as the input
argument, and that the file type is defined before the input argument.

I think that is the same issue as
https://discourse.llvm.org/t/using-x-c-with-clang-parsetranslationunit/18927/5

In order to do that, we now filter the task commands of the input
argument, to manually add it later on.

For dependencies, we also filter the output argument and use
'-o/dev/null' to avoid any issues.

Change-Id: I33a585be85e0c7962690d19ba0d6d7784d86713e
Priv-Id: 69eaec39317cf8019cc08dc6b6b2f74b2e76f7ca