http: fix noactDelay handling after a Connection: close
The HTTPd noactDelay will close connections which present no activity
and for which *all* the pending queries have been answered.  Indeed, a
connection cannot be considered as inactive if some queries are awaiting
an answer from us.

However, when a client set the `Connection: close` header to tell it
wants to close the connection *after* the reception of the query answer,
we enter a special case in httpd_on_event() which didn't disable the
activity watcher in case of pending queries.

So this patch rework the code in order to handle the state
HTTP_PARSER_CLOSE in the same way we handle the state HTTP_PARSER_IDLE.

Change-Id: I5a5416e21c7ce18c93fe72396cca4a3a5028dc87
Priv-Id: febf80ee38317478ca9759c61b24b5c5243db0d1