qps: ensure correct map generation value on map metadata changes
* What
This patch handles cases where metadata in QPS page maps are actually
modified but data content of there maps are not yet changed, leading to
invalid QPS snapshots.

* Why
When functions qps_pg_map_int(), qps_pg_remap_int() or
qps_pg_unmap_int() are called after a snapshot, it is possible to
unsynchronize QPS page map metadata and its content. This is triggered
by *NOT* backing up this map if there is no write operation on it in the
meantime. This behavior can lead to:
 - memory leak in QPS page maps (reload of this map from a snapshot +
   restart will keep the pages as allocated and definitely lost)
 - a memory block allocated but not yet accessed can be allocated for
   someone else too (after snapshot + restart).

* How
By setting the current QPS generation value on the map when an update of
map generation value is necessary.

Change-Id: Ie2c88ba335b669e4694509b937610f80c73cc34d
Priv-Id: f0777516fa2c7dad91b0b8c1dd464abb4bf7a965