z: in 'z_diff()', make the expectations clearer
The function 'z_diff()' is used in 'z_iop_diff()', that is called in
case of failure when comparing IOPs.

Rename the diff files we use to generate the diff in 'z_diff()' so that
the result we have and the result we expect are easier to identify
("exp" and "got" instead of "left" and "right").

Change-Id: I9c888cbd36229e4982e54e167bc031f48b05d353
Priv-Id: 756920026b58ea3393c4bf5a35a580511687f0ce