build/asdf: add template support for ASDF tool-versions file.
ASDF tool-versions does not support conditions.

In our case, we want to use the system version if the Python >= 3.9.
Else, we want to use Python 3.9.

In order to this, we now use a templating file `.tool-versions-tpl.sh`
that is executed to generate the final `.tool-versions`.

The python system version is computed by the script
`asdf_python_version.sh` that is run by the template file.

It is possible to manually set the different environment variables
TOOL_VERSIONS_PYTHON, TOOL_VERSIONS_DEFAULT_PYTHON,
TOOL_VERSIONS_MIN_PYTHON and TOOL_VERSIONS_MAX_PYTHON for some
customization of .tool-versions.

`asdf shell python ...` and `$ASDF_PYTHON_VERSION` still override the
configured version in .tool-versions.

Change-Id: I559b8fdd8efe71d36866405df1f74de26a33ebd6
Priv-Id: 3e27bcee4c9569a354f24c6a6710f02093523f18