wscript: rework ASDF and poetry waf configuration.
In the next commits, we will had the possibility to handle poetry with
ASDF, and to use the system python instead of the selected version
with ASDF.

This adds a new complexity when switching branches where the previous
version of python could have been instead of system one because of the
way poetry and ASDF are modifying the PATH when running a python script
(which waf is).

We need to make sure to have a clean PATH of all previous python
environment when using ASDF and poetry before creating the new poetry
environment.

Move some the ASDF functions in a separate fold to be more readable.

Change-Id: Icbcb1cfabd3ed854127fd0aba09dbe278610ae97
Priv-Id: 5e43e33258a23156e3732437d7e40839f1c2a005