iopy/cython: update Cython to 3.0.8 and fix IOPy for this version.
Since Cython 3.0.x, cdef functions are no longer assumed to be
'noexcept'. It needs to be put manually and is even required in some
cases.
This means that the exceptions are now checked even for plain C
functions that don't manipulate Python objects.
As a result, it is recommended to now use `noexcept nogil` when possible
on plain C functions that don't manipulate Python objects.

See the migration guide
https://cython.readthedocs.io/en/latest/src/userguide/migrating_to_cy30.html

Due to https://github.com/cython/cython/issues/6014, we need to ignore
one more error when compiling some Cython files.

Change-Id: I733e905ffa140f31835878e5f72884f0cd6bdf3a
Priv-Id: 1ec8318fb2c74178ee1560f5b6d3d090b04b2fc5