qps/hat: fix invalid node access preceding empty leaf
In qhat_optimize_parent() function, it was possible to reference an
invalid leaf when the current leaf has just been emptied, leading to
this assertion failure:

qhat_unflatten_leaf64: Assertion `old_node.leaf' failed.

During the search of the leaf to optimize, the number of entries in
leaves is retrieved with variables before_count and after_count.
However, the algorithm did not check that before_count was actually set
before referencing a leaf to optimize (the current logic
"before_count <= after_count" is true even if there is no item in the
previous leaf or no leaf at all).

Fix this issue to ensure a legitimate leaf later by ensuring
that before_count variable is set (i.e. a previous leaf exists and
contains at least one item).

Change-Id: Ifbf3bab5f8c94406f1638182fee0b54d1157eb05
Priv-Id: 3a0f6e5ac7d9d3a34dba806074527170b9c5ba05