net/https: set SNI for TLS HTTP connections.
SNI is an optional field that can be set by clients when they connect to
a server with HTTP protocol.

Add a field in HttpcCfg configuration to set it for connections.
The value can also be set by the user in httpc_cfg_t to be more
configurable (for example, if product has to be connected to different
servers that have different SNI).

Change-Id: Id45c16466ce4e188a11b3701b91679d2ccd40add
Priv-Id: 53eae63ba9ed6948c5545b255ee82f0a6246c786