qps/bitmap: fix qbitmap double free on QPS pages
The current code in qps_bitmap_clear() doesn't properly wipe the
allocated QPS pages when qps_pg_unmap() has been performed on nodes. As
a result, it is possible to call again this function or
qps_bitmap_destroy() and corrupt the QPS allocator (please see the end
of this commit for more explanations). This defect makes also possible
the reuse of QPS bitmap nodes after they have been freed.

This bug has also shown that no protection preventing an invalid double
free for QPS pages was put in place. Based on QPS block metadata, it
can easily be checked by triggering a panic, before corrupting the QPS
allocator.

In full details, the QPS allocator can effectively be corrupted by
replaying the following sequence before this patch:
 - Create QPS bitmap and set keys like in nr_100747 test
 - Call qps_bitmap_clear() twice
 - Call __qps_check_maps()

The assert triggered by qps_pg_check_hdrs_aux() is this one:
    {qps/bitmap} hdrs[pg + sz].flags & QPS_BLK_PREV_FREE

When the issue occurs, we read an invalid size metadata from a previous
free/used block in the past but not anymore. This invalid read changes
the chained list for the next blocks to reach, potentially jumping on
another invalid metadata... or no metadata at all.

Change-Id: Ie19246d6675522f13781d3f65603e9ce0de7299c
Priv-Id: 5ad3e62d1014180302bda8f2c2ed69d2f0dd5a3c