qps/hat/bitmap: remove unused QPS unload functions
Functions to unload memory pages are currently implemented by QPS. This
could be not problematic if the equivalent functions were implemented
to load back memory in QPS, before performing memory accesses. This is
unfortunately not the case: unloading memory in QPS can only be restored
through a full reopen of QPS to get it back.

As these functions are not used at all, it has been decided to remove
them.

Change-Id: I47a8271e73dd518371b637fa10aa2d091ecbee25
Priv-Id: 7b27ec9dc7eea80f19a41eabbfcb570d1c6b1778