qps/hat: fix invalid node access following empty leaf
In qhat_optimize_parent() function, it was possible to reference an
invalid leaf when the current leaf has just been emptied, leading to
this assertion failure:

qhat_unflatten_leaf64: Assertion `old_node.leaf' failed.

During the search of the leaf to optimize, a search on nodes after the
current emptied leaf is performed. Unfortunately, when a leaf is
eligible, the current algorithm doesn't take the leaf at the index but
the one which is before this index (probably an invalid copy/paste
from before_count?). This is wrong and even not coherent with the
affectation of after_count which is retrieved from the right leaf, few
lines before.

Fix the index on after_idx variable, so a legitimate leaf is used later
on.

Change-Id: I7a19aac7673403d0da4ea93f6d75804923e27bf6
Priv-Id: ae877e6135b8017cb1b206d6978bf3373c910d52