core: fix file reading
Squash commit of multiple commits:

---

core: fix file reading from virtual FS

The lib-common provides two set of APIs to read a complete file:
- lstr_init_from_{file,fd} which aims to mmap() a file through a lstr_t
  for convenience;
- sb_read_{file,fd} which aims to efficiently read() a file by
  pre-allocating the needed memory and performing a “read at once” when
  possible.

Since the purpose of these APIs is to easily read any file at once, they
also support reading from pipes, fifos and virtual FS (/proc, /sys,
etc…). Such a file was detected by these API based on the fact that
stat() would return a `st_size` of zero, which is often true but not
*always* true…

For instance, most of the files in /proc reports a size of 0, like
/proc/diskstats:

  $ stat /proc/diskstats
    File: /proc/diskstats
    Size: 0    Blocks: 0    IO Block: 1024    regular empty file

But, if /proc files are exposed through a fused for example, like on a
LXC using LXCFS (see
https://stgraber.org/2016/03/31/lxcfs-2-0-has-been-released/) the size
of the file will be shown:

  $ mount | grep diskstats
    lxcfs on /proc/diskstats type fuse.lxcfs (rw,nosuid,nodev,relatime,
      user_id=0,group_id=0,allow_other)
  $ stat /proc/diskstats
    File: ‘/proc/diskstats’
    Size: 1229    Blocks: 0    IO Block: 4096    regular file

In this case, sb_read_{file,fd} actually succeed being read-based.
However lstr_init_from_{file,fd} failed because it tried to mmap() the
file, but mmap() being unsupported on a virtual FS like /proc, it failed
with a ENODEV error.

In addition, a third case is possible. A file served by a virtual
file-system can also be shown with a default size which doesn't match
its true content. This is the case of most of the files served under
/sys/kernel:

  $ stat /sys/kernel/profiling
    File: /sys/kernel/profiling
    Size: 4096    Blocks: 0    IO Block: 4096    regular file
  $ cat /sys/kernel/profiling
  0

In which case we can only use the reported size as a hint but we need to
rely on read() to detect the end of the file. sb_read_{file,fd} failed
on this case too as it expected to match the exact st.st_size when
it was greater than zero.

So basically this patch changes our APIs to handle of these
possibilities correctly.

Change-Id: Icad4b53a6dc3a1c674441f8968610631d33ad288
Priv-Id: fc1f06b06aa5e3369ff89ea4abf068869602a014

---

core/lstr: don't use read instead of mmap for shared writing

The lstr_init_from_{file,fd} allows to mmap a file for both reading or
writing. When PROT_WRITE is asked, we must not allow to fallback on
read() when mmap() is not available and the mmap isn't private.

Change-Id: I31418eb800893d5dc5859bda7a4351fd531acce2
Priv-Id: b9ec34ccbd04bda99bb6ad32ca3a292d34162ec0

---

str-buf: fix buffer pre-allocation in sb_read_fd

In fc1f06b06aa5e3369ff89ea4abf068869602a014 we used sb_grow() to
pre-allocate the expected file size and thus optimize the following
sb_read() calls. However, sb_read() would call in turn sb_grow() with a
hint of BUFSIZ which could shrink the allocation because of the
sb_optimize() called in sb_grow().

To fix this we compute the right hint in sb_read_fd() and we always
provide it to sb_read().

Change-Id: Ib50b60ae93408cbd9b3ba7d0447c3307082d6680
Priv-Id: 163a1280d865b0b247ebad630d3ce7ff00d4d204

---