qps/hat/bitmap: remap key and end flag for nullable Qhat
The current Qhat enumerator has duplicated variables for nullable Qhats,
which are:
 - Key value
 - End flag when the enumerator is done.

When the QPS bitmap enumerator has to be synchronized again with the
bitmap (i.e. deletion of keys, for example), the main enumerator for
nullable Qhats can be totally out of sync, leading to a general
inconsistency.

As the variables of both QPS bitmap enumerator and Qhat enumerator have
similarities, it is possible to keep a public access on the right key
variable and end flag by moving variables in QPS bitmap enumerator,
while still accessing structures dedicated to each fonctionnal Qhat.

Some static assert have been added to certify that memory location is
used as expected and developpers notified on any invalid changes in an
enumerator structure.

This change allows us to remove the calls to 'qhat_enumerator_catchup()'
from 'qhat_get_enumerator_at()', 'qhat_enumerator_next()' and
'qhat_enumerator_go_to()': in these cases, the call only synchronized
the key attribute, but it is always synchronized now.

Change-Id: Ie49a38248f8b36f47df5e2d5a266cf88e4ae55f5
Priv-Id: ab6a118c79c707fe050bbd274e460cbbd0cee35d