qpsstress: stop at the end of a replay
* What
Improve the replay feature of qpsstress by stopping the application at
the end of a replay sequence (no random events are generated
afterwards).

* Why
It is annoying sometimes to just check if a fix on QPS has been
performed successfully but not being able to see the operations in a
terminal because random events are systematically generated afterwards.

* How
By extending the replay mechanism and allow the application to stop when
replay has been executed successfully.

Change-Id: I23c589d5fe151c2efb9e3dc38ef6b9ef6bea049b
Priv-Id: 159fbe1f8cde87f74386ee3302f673da9d96691f