qpsstress: change log management
* What
Improve the way logs are written for each step in qpsstress, so we can
know which step triggers the assert or segfault.

* Why
When an issue occurred, the last step was not written in log and had to
be retrieved from the core dump or by guessing the callstack from the
assert. Even if it is not always possible to display all information
before playing a step (especially during allocation/reallocation of
objects), we should easily get more information from this tool.

* How
By moving some logs and by changing the level of logs before executing
step when all information is not known in advance (so it can still be
easily replayed by the tool with minimal modifications).

Change-Id: Ic852ca13e63f4ab385aeef33efa335404e5500f8
Priv-Id: 812e7486705cf2a10e68d8e5a3f35f7e0dea6be4