qps: change qps_pg_check_hdrs_aux() parameter
Prior to this patch, the second parameter named 'blk' (a QPS page id in
the map) was used to move on the first qps_pghdr_t *hdrs of the QPS page
map to check. This was obviously wrong as all callers used the map index
of QPS with a QPS_MAP_PAGES shift to convert it as a block in the map.

Fix it by explicitly asking for the map index, and change the prototype
name so the signature of this function is not badly used with old code
providing a block id instead of a map index.

Change-Id: I062df560dffd17d47708d3050ad815c5eb8e8bc4
Priv-Id: e1b255ffa7bf896a03b1b9188b58d170daa9846d