iop-dso: add the support of a custom user version
The IOP DSO are now exported with an optional custom user version and a
verification callback which are exposed in the iop_dso_t object.

This will allow the users of the IOP DSO (Python, PHP, etc…) to provide
this user version to the ichannel library and thus pass the backward
compatibilities check implemented in the platform's products.

Change-Id: I698b7f08f66c6836ffd618ac2399223d22cdf2c8
Priv-Id: 53df29b6233596cd763b3552d0aa156a50fe4169