ichannel: disable user version checks for unix sockets
When using Unix sockets in the IChannel library, all the handshake
exchanges are bypassed because we assumed that two services establishing
a Unix socket are necessarily fully compatible.

If this indeed allow faster connections establishment, it also prevents
to do some useful checks like version compatibility. It also does weird
things, like setting the ic->peer_version to 0 instead of IC_VERSION but
only on server side.

Thus for now we disable the user version checks with Unix sockets, but
we should probably rework their connection process to allow a lighter
handshake or something (like always enqueuing the version message
instead of a nop message without actually blocking to perform the full
handshake).

Change-Id: I499d0ff476d37c94aad6bbd1a2b172ba20e38ccb
Priv-Id: bf7a9345466f735b39275352f9f5f727d284f87b