qpsstress: improve Qhat enumerator on trie changes
* What
Improve the way a Qhat enumerator is managed, so we can have strong
checks on safe/unsafe operations around this enumerator (i.e. only safe
calls are allowed until a safe next move on enumerator on Qhat changes)
or weak checks (i.e. the enumerator is dumb and can randomly
re-synchronized partially after few safe calls and be unsafe before
moving enumerator).

* Why
Before this patch, qpsstress didn't manage properly the Qhat enumerator
in specific cases where some keys were randomly removed after
positioning the enumerator but before reading it. As a consequence, the
Qhat enumerator could be at an invalid position or even re-synchronized
partially, despite safe accesses, leading to a segfault.

Even if the Qhat issue has to be resolved in a dedicated patch, this
tool should be able to give the choice to enforce enumerator accesses.
In fact, changing the Qhat structure should only trigger safe accesses
until we move safely to the correct next enumerator, as this is unlikely
to happen in the real world.

* How
By providing an option for this tool and a context for each
enumerator when structural changes occur on the associated Qhat.

Change-Id: If8d62e5d4befc972df7bcfa05c9a9be5da85a3fd
Priv-Id: aa78e8f5268374a477607d92a5677e1c6e6c6d4d