qps/hat: fix assertion when path is synchronized but not the enumerator
When the enumerator tries to read a key which has just been removed, the
path stored inside this enumerator is re-synchronized and the key is
marked as removed (aka en->key_was_removed = true). However, in order to
move forward, the en->key is set back to the removed key, assuming
a forward move later will set en->key to the proper available key in the
trie.

This is unfortunately not true for the qhat_tree_enumerator_go_to() call
which trigger this assertion even in safe access if the key is above the
removed key but still below the en->path.key:

    qhat_tree_enumerator_find_forward: Assertion `key >= en->path.key'
    failed

In fact, the enumerator is not able to move to the new location provided
in path structure of the enumerator which is already synchronized.

Fix this issue by getting rid of the attribute
'qhat_tree_enumerator_t::key_was_removed' introduced by patch
f3a797e09a9 which handles very specific cases but adds also complexity
in 'next' and 'go_to' functions. Now, we do it the hard way: the next
call to 'next' or 'go_to' will need a full lookup. This is suboptimal
but we do not care: these specific cases when the user calls
'qhat_tree_enumerator_get_value()' after removing the current key is not
supposed to be usual.

Change-Id: I5513996b2d354bd0f18e715d84a2966f9d53aed7
Priv-Id: 3d35abe7ca55de7d69d80ed6bd28ad0760967a5d