iop: remove global IOP environment and pass it down to the functions.
Instead of having a global IOP environment variable, have the user
creates a custom IOP environment and pass it down to the different IOP
functions.

This totally breaks the compilation and tests in lib-common.
But to avoid having one really big modification commit, this will be
solved file-by-file in the next commits.

Change-Id: I64e9d8e81b0c4e60d725bb0f523098696aa09220
Priv-Id: 6385f886732137fc13c19fbd15112960c8c442e6